src/client/requirements_client.txt
numpy
pyserial
Pillow
matplotlib